Complete Haiku functionality
After a user completes a haiku, 1) he should see his haiku in a form,
with all fields editable, and a message that says "You can edit your
haiku before you complete it". Allow that user to edit any line in that
haiku -- even a friend's line, if he wrote it with a friend. Supply a
button that says "Submit", and when the user clicks that button, save
the haiku, and show a flash message that says "Your haiku is completed!"